consolidated ProductQuestion helpers
