--- client/src/components/product-questions/ProductQuestionHelpers/getNewAnswer.js
@@ -1,12 +1,0 @@
-const getNewAnswer = (questionList, questionId, newAnswers) => {
-  const newList = questionList.map((question) => {
-    if (question.question_id === questionId) {
-      question.answers = newAnswers;
-    }
-    return question;
-  });
-
-  return newList;
-};
-
-export default getNewAnswer;
