--- client/src/components/product-questions/ProductQuestionHelpers/getQuestionHelpful.js
@@ -1,12 +1,0 @@
-const getQuestionHelpful = (questionList, id) => {
-  const newList = questionList.map((question) => {
-    console.log(question);
-    if (question.question_id === id) {
-      question.question_helpfulness++;
-    }
-    return question;
-  });
-  return newList;
-};
-
-export default getQuestionHelpful;
